Updated to twitch-irc 4.0.0
Alongside updating to the new version,
this also adds handling for joining incorrect channels,
which panics if the channel you're trying to join is not correct.

This is of course a bad thing, but since my bot is currently
single channel, I do not necessarily see this as a bad thing, as I will
need to fix it in the future anyway.

Signed-off-by: Mykola <ks2225@protonmail.com>